moving index info
