--- src/assets/images/index.ts
@@ -1,5 +1,0 @@
-export { default as i6 } from './i6.svg';
-export { default as appStore } from './appstore.svg';
-export { default as play } from './play.svg';
-export { default as smilies } from './smilies.svg';
-export { default as social } from './social.svg';
