+++ src/assets/index.ts
@@ -1,0 +1,5 @@
+export { default as i6 } from './images/i6.svg';
+export { default as appStore } from './images/appstore.svg';
+export { default as play } from './images/play.svg';
+export { default as smilies } from './images/smilies.svg';
+export { default as social } from './images/social.svg';
