+++ src/components/common/Button.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 
-interface ButtonProps extends React.ButtonHTMLAttributes<any> {
+interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   children: React.ReactNode;
 }
 
-const BlackButton = ({ children, ...props }: ButtonProps) => (
+const Button = ({ children, ...props }: ButtonProps) => (
   // eslint-disable-next-line react/button-has-type
   <button className="blackButton" {...props}>
     {children}
   </button>
 );
 
-export default BlackButton;
+export default Button;
