--- src/components/common/index.ts
@@ -1,4 +1,0 @@
-export { default as Button } from './Button';
-export { default as InfoSection } from './InfoSection';
-export { default as InputField } from './InputField';
-export { default as PageWrapper } from './PageWrapper';
