+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from './common/Button';
+export { default as InfoSection } from './common/InfoSection';
+export { default as InputField } from './common/InputField';
+export { default as PageWrapper } from './common/PageWrapper';
